refactor: move files to folders
