+++ first-react-app/src/app/App.tsx
@@ -1,10 +1,10 @@
 import "./App.css";
 import React from "react";
 import { HashRouter, Route, Routes } from "react-router-dom";
-import Home from "./pages/Home";
-import NotFound from "./pages/NotFound";
-import About from "./pages/About";
-import Forms from "./pages/Forms";
+import Home from "../pages/Home";
+import NotFound from "../pages/NotFound";
+import About from "../pages/About";
+import Forms from "../pages/Forms";
 
 localStorage.setItem("searchValue", "");
