+++ first-react-app/src/components/formComponents/FormCard.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import note from "../assets/note.png";
+import note from "../../assets/note.png";
 
 function FormCard(props: {
   songName: string;
