+++ first-react-app/src/components/homeComponents/Card.tsx
