+++ first-react-app/src/components/homeComponents/CurrentPage.tsx
