+++ first-react-app/src/components/homeComponents/SearchBar.tsx
